Cleancut/requirements.txt
fastapi
uvicorn[standard]
gunicorn
rembg[cpu]
stripe
python-dotenv
python-multipart
